compaction: allow controller maximum shares to be configured
A new config option compaction-max-shares is introduced to allow
configuration of maximum controller output (shares).

The motivation is to allow user to change the upper bound in the
range of shares the controller works with. The default can be
lowered to reduce compaction aggressiveness. And likewise, the
default can be increased to help compaction catch up with the
backlog.

Controller uses a linear function to map normalized backlog to
shares, so the slope is being preserved regardless of the
maximum shares configuration. So only the limit is being
increased without changing the function behavior.
Subsequent work will introduce an aggressiveness factor that
will allow us to control the slope (which reflects the
normalization factor).

            Max shares
Backlog     500         1000        2000
         |=================================
1        |  83.3333     83.3333     83.3333
2        |  115.789     115.789     115.789
4        |  178.947     178.947     178.947
8        |  305.263     305.263     305.263
16       |  500         557.895     557.895
32       |  500         1000        1063.16
64       |  500         1000        2000

Fixes #9431.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>