raft topology: Always set host id in topology
Before, it was updated only for normal nodes. We need it for
bootstrapping nodes too.  Otherwise, algorithms, e.g. load balancer,
will be confused by observing nodes in topology without host id set.

This will become a problem when load balancer is invoked concurrently
with bootstrap, which currently is not the case, but will be after
later patches.

We should maintain that all nodes in topology have a host id.