sstables: switch entry_descriptor(sstring..) to std::string_view
so its callers don't need to construct a temporary `sstring` if
the parameter's type is not `sstring`. for instance, before
this change, `entry_descriptor::make_descriptor(const std::filesystem::path...)`
would have to construct two temporary instances of `sstring`
for calling this function.

after this change, it does not have to do so.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>