open-coredump: do not assume remote repo name
open-coredump.sh allows us to specify --scylla-repo-path, but
developer's remote repo name is not always "origin" -- the
"origin" could be his/her own remote repo. not the one from which
we want to pull from.

so, in this change, assuming that the remote repo to be pulled
from has been added to the local repo, we query the local repo
for its name and pull using that name instead.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes #15220