toppartitions: Add support for probabilistic sampling to data_listeners
This patch adds to the data_listener class a sampling_probability parameter which
allows us to control the probability of sampling a query, mimicking what's done
for probabilistic tracing in tracing/tracing.cc.
This allows us to be able to mitigate the performance impact when a data_listener
is being used.
The sampling_probability must be a double between [0, 1], with 0 meaning that
the data_listener is disabled and 1 meaning that 100% of the queries will be sampled.