toolchain: support building an optimized clang
This is a different way attempting to combine building an optimized clang (using LTO, PGO and BOLT, based on
compiling ScyllaDB) to dbuild. Per Avi's request, there are 3 options: skip this phase (which is the current default),
build it and build + install it to the default path.

Right now, the prepare script is hard-coded to skip this phase. A follow-up patch could make it configurable.

Fixes: #10985
Signed-off-by: Yaniv Kaul ykaul@scylladb.com