system_keyspace: remove redundant wait_for_sync_to_commitlog
Tables with schema commitlog already sync every
write, wait_for_sync_to_commitlog makes sense
only for the regular commitlog.

Technically there are nothing wrong with allowing
both options, but it's confusing. Being strict
and accurate about the meaning of the options
reduces the chance of errors due to misunderstanding.

This is preparation for the next commits, where
we will start generating an error if the combination
of options doesn't make sense.