db/hints: Extract the logger used in manager.cc
This is the initial step in the process of modularization of manager.cc.
The file will be divided into smaller units to improve readability,
better isolate certain components, and finally make reasoning about
the code and control flow better.

To preserve (at least for the time being) the same logging behavior,
i.e. the same logger name, we extract the logger to a separate header.
It will be included in all files originally making for manager.cc.

It will most likely change in the future, and individual loggers will
be used for each of the files instead.