db/hints: Extract sender from manager.hh
This is the first step in the modularization of the module.
All relevant data structures should be handled in separate translation units
to improve readability and maintainability. It should also help in isolating code.

This commit extracts the data structure, which is the lowest in the dependency
hierarchy in manager.hh -- sender -- responsible for sending hints to an endpoint.