db/hints: Extract host_manager from manager.hh
This is the second step in the modularization of the manager.hh.
We move the logic of end_point_hints_manager to a separate file
to improve readability, better isolate the code, and finally
help maintaining the module by splitting it into smaller units,
which are easier to reason about.

This commit also renames the data structure to host_manager, which
is shorter while still conveying its purpose.