db/hints: Extract the hint rebalancing logic from manager.cc
This commit is another step in modularization of the Hinted Handoff module.

This commit moves the logic related to rebalancing hints at Scylla's start-up
to another file. The purpose of these changes is to isolate functionalities
not directly linked to the purpose of manager.hh -- the class declared in this
file deals with managing hints for a specific shard. However, rebalancing
hints at start-up moves them BETWEEN shards.

What's more, the only call to the function comes from directory_initializer,
which, though declared in the same file, has another purpose, so keeping
the functionality as part of hint shard manager makes little sense.

The commit also renames the function exposed to the user from 'rebalance'
to 'rebalance_hints' to better convey what it does.