db/hints: Reorder functions in host_manager.cc
This commit reorders functions in the source file to match the order
in which they are declared in the header file. This should make it
easier to find something without resorting to features of editors.
The order in the header file also makes sure that related functions
are close to each other.