repair: Use full endpoint location instead of datacenter part
There are several places in repair code that get datacenter from the
topology. Nearby there are calls to update_topology() which, in turn,
needs full location ({dc, rack} pair). This patch makes the former
places obtain full location from topology and get the dc part from it.
This is needed as a preparation to let latter places use that location.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>