db/hints: Remove an inline specifier of have_ep_manager
The inline specifier leads to internal linkage, i.e. though the function
is declared in the header file, it cannot be accessed from other
translation units.

On one hand, that is not a problem as the function is private, but it
still makes no sense to do that -- the compiler will most likely inline
it in the source file anyway and in general, it is a mistake.