db/hints/manager.hh: Ditch a never implemented member function
The function make_directory_initializer is not implemented or referenced.
The current version of Scylla seems to create instances of directory_initializer
explicitly, i.e. by calling directory_initializer::make() rather than using
a hint shard manager (simply called manager in the code) for that. Removing it
has no effect on the codebase, while keeping it may cause confusion to the reader
not familiar with the module or its usage.