db/hints: Introduce a member function version of with_file_update_mutex
This member function is going to replace the existing friend implementation.
The purpose of this transition is to better isolate the code -- no one except for
hint_sender and manager should know about the existence of host_manager.

As of now, resource_manager is one example of a piece of code that tries to use
host_manager directly. We want to change that.