compaction: allow configuration of controller's backlog sensitivity
A new option compaction-backlog-sensitivity is introduced to
allow configuration of system's sensitivity to compaction backlog
using controller's normalization factor (so this one is indirectly
configured through it).

Some background:

Normalization factor (NF) controls the trade-off between reads
and writes. The lower the NF, the higher the aggressiveness.
A suboptimal NF leads to compaction either falling behind or
running more aggressive than needed.

Exposing raw NF is not good because it's an implementation detail,
and hard to reason about. So we're introducing a concept of
backlog sensitivity (BS) that abstracts NF. Another way to think
of sensitivity is weight assigned to each backlog.

The controller uses a linear function for mapping backlog to
shares, and uses interpolation for finding shares given a backlog
between minimum and maximum shares.

The formula for finding the slope is:
    m = (max_shares - min_shares) / (NF - min_shares_backlog)
So NF plays a large role in the steepness of the slope.

To find the NF given any number of shares, we do:
    NF = (max_shares – min_shares) / m + min_shares_backlog

The BS is applied to the slope, so a BS of 2.0 means that
NF is lower, and therefore the backlog sensitivity is higher.
And also the other way around.

See:

            Backlog sensitivity
Backlog     0.5         1.0         2.0
            |=================================
1        |  83.3333     83.3333     83.3333
2        |  107.895     115.789     131.579
4        |  139.474     178.947     257.895
8        |  202.632     305.263     510.526
16       |  328.947     557.895     1000
32       |  581.579     1000        1000
64       |  1000        1000        1000

In the future, we can tune sensitivity automatically based on
known internal variables (disk:ram ratio, write amplification,
available write bandwidth, available disk space, etc).

Refs #15096.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>