doc: install headless jdk in sample command line
in comparison with java-11-openjdk, java-11-openjdk-headless does not
offer audio and video support, and has less dependencies. for instance,
java-11-openjdk depends on the X11 libraries, and it also provides
icons representing JDK. but since scylla is a server side application,
we don't expect user to run a desktop on it. so there is no need to
support audio and video.

in this change, we just suggest the a "smaller" package, which is
actually also a dependency of java-11-open-jdk.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>