doc: run install.sh directly
strictly speaking, `sh` is not necessarily bash. while `install.sh`
is written in the Bash dialect. and it errors out if it is not executed
with Bash. and we don't need to add "-x" when running the script, if
we have to, we should add it in `install.sh` not ask user to add this
option. also, `install.sh` is executable with a shebang line using
bash, so we can just execute it.

so, in this change, we just launch this script in the command line
sample.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>