SCYLLA-VERSION-GEN: do not print version by default
actually, we never use the its output in our workflow. and the
output is distracting when building the package. so, in this
change, let's print it only on demand. this feature is preserved
just in case some of us would want to use this script for getting
the version number string.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>